Fixed bug where TileReactorCell was reading float instead of double causing client side temperature to be wrong.

+++ src/main/java/org/halvors/nuclearphysics/common/tile/reactor/TileReactorCell.java
@@ -218,7 +218,7 @@ public class TileReactorCell extends TileInventory implements IFluidHandler, IRe
         super.handlePacketData(dataStream);
 
         if (worldObj.isRemote) {
-            temperature = dataStream.readFloat();
+            temperature = dataStream.readDouble();
             tank.handlePacketData(dataStream);
         }
     }
